node: Upgrade to node 18 (+ everything else)
This change upgrades node to `v18` which also includes upgrading `npm`,
the `package-lock.json` and `node2nix`.

One issue with this upgrade was switching out the node version caused
building the frontend shell to fail due to incompatibilities between the
existing `npm` version used by the `node2nix` files and the
`package-lock.json` file. It tried to migrate the outdated
`package-lock.json`, but that failed because it runs without network
access and I only assume it then didn't have the appropriate hashes for
all the packages and simply failed.