app: Ask for PIN on join a session
This change adds the requirement to provide a PIN to join an existing
session.

The PIN is generated by the server when creating a new session and shown
to the session creator.

This change also changes the ConnectionManager to use a Context object
to provide ControlSocket and PeerConnection objects to contained
compontents instead of some ad-hoc `useState` prop drilling.