feat: Add "Link" share item type
This change adds a "link" share type and moves the "hidden" state into a
separate attribute. This allows for hidden links and regular text. The
share item type is automatically set to the "link" type if the input
looks like an URL based on some very basic heuristics. Changing the
share item type manually disables the automatic type selection until the
next share.