feat: Add client session + client IDs
This adds support for tracking clients through an encrypted JSON token
stored as a client-side cookie. The server assigns a client ID to each
client and persists this with the client. If the client clears cookies a
new client ID is generated.

The plan is to also include authentication data in this token to be able
to authenticate with the TURN server. Also it's useful for rate-limiting
and tracking bandwidth used through the TURN server.